[rocon_python_comms] extracted ConnectionCacheProxy to another file. set diff publisher queue to 5 to avoid losing messages.
