+++ rocon_python_comms/src/rocon_python_comms/__init__.py
@@ -18,8 +18,9 @@ ros python communications.
 # Imports
 ##############################################################################
 
-from .connections import Connection, ConnectionCache, ConnectionCacheNode, ConnectionCacheProxy
+from .connections import Connection, ConnectionCache, ConnectionCacheNode
 from .connections import SUBSCRIBER, PUBLISHER, SERVICE, ACTION_CLIENT, ACTION_SERVER, connection_types
+from .connections_proxy import ConnectionCacheProxy
 from .exceptions import *
 from . import master
 from .nodes import find_node
